move debugger package to packages dir
