+++ packages/debugger/deno.json
@@ -2,8 +2,8 @@
   "name": "@fedify/debugger",
   "version": "1.9.0",
   "exports": {
-    ".": "./mod.ts",
-    "./cli": "./cli.ts"
+    ".": "./src/mod.ts",
+    "./cli": "./src/cli.ts"
   },
   "publish": {
     "include": [
@@ -18,7 +18,9 @@
     ]
   },
   "imports": {
-    "@fedify/fedify": "../fedify/mod.ts",
+    "@fedify/fedify": "../fedify/src/mod.ts",
+    "@fedify/fedify/federation": "../fedify/src/federation/mod.ts",
+    "@fedify/fedify/vocab": "../fedify/src/vocab/mod.ts",
     "hono": "jsr:@hono/hono@^4.8.3",
     "@std/assert": "jsr:@std/assert@^0.226.0",
     "@std/async": "jsr:@std/async@^1.0.0",
