+++ packages/debugger/tsdown.config.ts
@@ -1,0 +1,17 @@
+import { defineConfig } from "tsdown";
+
+export default defineConfig({
+  entry: "src/mod.ts",
+  dts: true,
+  unbundle: true,
+  platform: "node",
+  external: [
+    "@fedify/fedify",
+    "@fedify/fedify/federation",
+    "@fedify/fedify/nodeinfo",
+    "@fedify/fedify/runtime",
+    "@fedify/fedify/vocab",
+    "@fedify/fedify/webfinger",
+    "@hono/hono",
+  ],
+});
